Removed insecure SSL configuration.
As it was, the openID auth was vulnurable to MITM attacks. Unfortunately is a common practice to NOT verify the host in curl, due to an old curl version that had outdated certificates. If anyone is still working on an out-dated curl version he should check out http://stackoverflow.com/questions/316099/cant-connect-to-https-site-using-curl-returns-0-length-content-instead-what-c for a secure fix.